Refactor text rendering
